+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ReleaseDate
+### Changed
+* Updated dependencies (cargo update)
+### Fixed
+* Some clippy lints
+
 ## [0.5.2] - 2019-09-12
 ### Changed
 * Switched to stable Rust